[update] add handlers

+++ BusinessLogic/Implement/BaseService/IBaseService.cs
@@ -2,11 +2,11 @@
 {
     public interface IBaseService<T> where T : class
     {
-        Task<T?> GetByIdAsync(int id);
+        Task<T?> GetByIdAsync(long id);
         Task<IEnumerable<T>> GetAllAsync();
         Task<T> AddAsync(T entity);
         Task<T> UpdateAsync(T entity);
-        Task<bool> DeleteAsync(int id);
+        Task<bool> DeleteAsync(long id);
     }
 
 }
