+++ Domain/Entities/Loan.cs
@@ -2,9 +2,9 @@
 {
     public class Loan : BaseEntity
     {
-        public int LoanID { get; set; }
-        public int MemberID { get; set; }
-        public int BookID { get; set; }
+        public long LoanID { get; set; }
+        public long MemberID { get; set; }
+        public long BookID { get; set; }
         public DateTime BorrowDate { get; set; } = DateTime.UtcNow;
         public DateTime? ReturnDate { get; set; }
         public User Member { get; set; } = null!;
